Timetable v2: Split into files
