+++ CODEOWNERS
@@ -1,2 +1,2 @@
 * @janzarsky
-src/components/TimetableV2.tsx @VojtaStanek @janzarsky
+/src/components/TimetableV2/ @VojtaStanek @janzarsky
